[ORC] Merge GetDylibInterface.h APIs into MachO.h.
These APIs are MachO specific, and the interfaces are about to be extended to
support more MachO-specific behavior. For now it makes sense to group them
with other MachO specific APIs in MachO.h.
